[CI]: Add shellscript check in lint stage ---
See merge request tikiwiki/tiki!4461

+++ doc/devtools/provide_tarballs.sh
@@ -26,43 +26,43 @@ TARGET="/var/www/tikiwiki.org/tar/"
 BASE=/opt
 # --- configuration --- edit here ---
 
-OLDIR=`pwd`
+OLDIR=$(pwd)
 PACKDIR=${BASE}/data/packaging
 
 mkdir -p ${BASE}/data/arc  ${PACKDIR}  ${TARGET}
 
 # start new logfile with timestamp as first entry
-date > ${LOGFILE}
+date > "${LOGFILE}"
 
 # always rebuild trunk from scratch (for now!)
 if [ "${BRANCH}" = "trunk" ]; then
-    rm -rf ${BASE}/data/${BRANCH}
+    rm -rf ${BASE}/data/"${BRANCH}"
 fi
 
 if [ ! -d "${BASE}/data/${BRANCH}" ]; then
     echo "Folder doesn't exist yet, checking out project ..."
 
-    cd ${BASE}/data
+    cd ${BASE}/data || exit
     if [ "${BRANCH}" = "trunk" ]; then
-        svn checkout --username ${SVNUSER} https://svn.code.sf.net/p/tikiwiki/code/${BRANCH} ${BRANCH} 2>&1 >> ${LOGFILE}
+        svn checkout --username "${SVNUSER}" "https://svn.code.sf.net/p/tikiwiki/code/${BRANCH}" "${BRANCH}" >> "${LOGFILE}" 2>&1
     else
-        svn checkout --username ${SVNUSER} https://svn.code.sf.net/p/tikiwiki/code/branches/${VERSION}.x ${BRANCH} 2>&1 >> ${LOGFILE}
+        svn checkout --username "${SVNUSER}" "https://svn.code.sf.net/p/tikiwiki/code/branches/${VERSION}.x" "${BRANCH}" >> "${LOGFILE}" 2>&1
     fi
 
-    LINE=`tail -1 ${LOGFILE}`
-    RC=`echo "${LINE}" | grep "Checked out revision"`
+    LINE=$(tail -1 "${LOGFILE}")
+    RC=$(echo "${LINE}" | grep "Checked out revision")
     if [ "${RC}" = "" ]; then
         echo "ERROR: could not check out project: ${LINE}"
         exit
     fi
 else
     echo "Folder exists, updating project ..."
-    cd ${BASE}/data/${BRANCH}
-    svn update 2>&1 >> ${LOGFILE}
-    LINE=`tail -1 ${LOGFILE}`
-    RC=`echo "${LINE}" | grep "At revision"`
+    cd ${BASE}/data/"${BRANCH} || exit" || exit
+    svn update >> "${LOGFILE}" 2>&1
+    LINE=$(tail -1 "${LOGFILE}")
+    RC=$(echo "${LINE}" | grep "At revision")
     if [ "${RC}" = "" ]; then
-        RC=`echo "${LINE}" | grep "Updated to revision"`
+        RC=$(echo "${LINE}" | grep "Updated to revision")
         if [ "${RC}" = "" ]; then
             echo "ERROR: could not update project: ${LINE}"
             exit
@@ -74,17 +74,18 @@ else
 fi
 
 echo "Cleanup packaging folder and rebuild"
-rm -rf ${PACKDIR}/${BRANCH}
-cp -r ${BASE}/data/${BRANCH} ${PACKDIR}
+rm -rf "${PACKDIR:?}/${BRANCH:?}"
+cp -r ${BASE}/data/"${BRANCH}" ${PACKDIR}
 
 # cleanup
 if [ -d "${PACKDIR}/${BRANCH}" ]; then
-    cd ${PACKDIR}/${BRANCH}
-    echo "Current folder: `pwd`"
+    cd ${PACKDIR}/"${BRANCH} || exit" || exit
+    echo "Current folder: $(pwd)"
 
     echo "Removing temp files and cache files, if any"
     # no tests needed and no temporary files:
-    rm -rf tests *.core *.tmp *.bak *.tpl.php temp/cache
+    rm -rf ./tests ./*.core ./*.tmp ./*.bak ./*.tpl.php ./temp/cache
+
     mkdir -p temp/cache
 
     echo "Hopefully no one has checked in his password file, we better remove it now"
@@ -97,46 +98,46 @@ if [ -d "${PACKDIR}/${BRANCH}" ]; then
     ./setup.sh -u www-data -g www-data -n fix
 
     echo "Removing svn config files"
-    find . -name ".svn" | xargs -n 1 rm -rf
+    find . -name ".svn" -print0 | xargs -0 -n 1 rm -rf
 
     echo "Adding timestamp file to archive"
     echo "Last update: " > .timestamp
-    echo `date` >> .timestamp
+    date >> .timestamp
 
-    cd ${PACKDIR}
+    cd ${PACKDIR} || exit
 
     echo "Creating tar.bz2 file"
-    rm -f ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2_previous
-    mv ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2 ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2_previous
-    tar -cjf ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2 ${BRANCH}
+    rm -f ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2_previous
+    mv ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2 ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2_previous
+    tar -cjf ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2 "${BRANCH}"
 
     echo "Createing zip file"
-    rm -f ${BASE}/data/arc/lastiki_${BRANCH}.zip_previous
-    mv ${BASE}/data/arc/lastiki_${BRANCH}.zip ${BASE}/data/arc/lastiki_${BRANCH}.zip_previous
-    zip -rq ${BASE}/data/arc/lastiki_${BRANCH}.zip ${BRANCH}
-    chown www-data:www-data ${BASE}/data/arc/lastiki_${BRANCH}.zip
-    chmod a+r ${BASE}/data/arc/lastiki_${BRANCH}.zip
+    rm -f ${BASE}/data/arc/lastiki_"${BRANCH}".zip_previous
+    mv ${BASE}/data/arc/lastiki_"${BRANCH}".zip ${BASE}/data/arc/lastiki_"${BRANCH}".zip_previous
+    zip -rq ${BASE}/data/arc/lastiki_"${BRANCH}".zip "${BRANCH}"
+    chown www-data:www-data ${BASE}/data/arc/lastiki_"${BRANCH}".zip
+    chmod a+r ${BASE}/data/arc/lastiki_"${BRANCH}".zip
 
     echo "Copying archives to docroot folder"
-    cp ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2_previous ${TARGET}
-    cp ${BASE}/data/arc/lastiki_${BRANCH}.tar.bz2 ${TARGET}
-    cp ${BASE}/data/arc/lastiki_${BRANCH}.zip_previous ${TARGET}
-    cp ${BASE}/data/arc/lastiki_${BRANCH}.zip ${TARGET}
+    cp ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2_previous ${TARGET}
+    cp ${BASE}/data/arc/lastiki_"${BRANCH}".tar.bz2 ${TARGET}
+    cp ${BASE}/data/arc/lastiki_"${BRANCH}".zip_previous ${TARGET}
+    cp ${BASE}/data/arc/lastiki_"${BRANCH}".zip ${TARGET}
 
     echo "Creating md5 files"
-    cd ${TARGET}
-    FILES=`/bin/ls * | grep -v md5 | grep -v grep`
+    cd ${TARGET} || exit
+    FILES=$(echo ./* | grep -v md5 | grep -v grep)
     for FILE in ${FILES}
     do
-      md5sum ${FILE} > ${FILE}.md5
-      chmod a+r *
+      md5sum "${FILE}" > "${FILE}".md5
+      chmod a+r ./*
     done
 else
     echo "ERROR: Folder not found after svn checkout."
     exit
 fi
 
-cd ${OLDIR}
+cd "${OLDIR} || exit" || exit
 
 echo "Finish"
